# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "New Voting System",
        "version": "1.0.0",
        "description": "API of the New Voting System. Conventions: docs/api/README.md."
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Same origin"
        }
    ],
    "paths": {
...
        "/v1/elections/{election}/ballots/order": {
            "put": {
                "operationId": "ballot.reorder",
                "description": "`ballots` must be exactly the UUIDs of the election's ballots, each once, in the wanted\norder (422 `ballots: set_mismatch` otherwise). A draft election only. Answers the list,\nin the new order. Owner or manager. Limited to 240 requests per hour per user.",
                "summary": "Put the ballots in order",
                "tags": [
                    "Ballot"
                ],
                "parameters": [
                    {
                        "name": "election",
                        "in": "path",
                        "required": true,
                        "description": "The election UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReorderBallotsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "position": {
                                                        "type": "integer"
                                                    },
                                                    "seats": {
                                                        "type": "integer"
                                                    },
                                                    "allow_blank": {
                                                        "type": "boolean"
                                                    },
                                                    "candidates_count": {
                                                        "type": "integer"
                                                    },
                                                    "candidates": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string"
                                                                },
                                                                "ballot": {
                                                                    "type": "string"
                                                                },
                                                                "party": {
                                                                    "type": [
                                                                        "string",
                                                                        "null"
                                                                    ]
                                                                },
                                                                "first_name": {
                                                                    "type": "string"
                                                                },
                                                                "last_name": {
                                                                    "type": "string"
                                                                },
                                                                "sex": {
                                                                    "type": "string"
                                                                },
                                                                "slogan": {
                                                                    "type": [
                                                                        "string",
                                                                        "null"
                                                                    ]
                                                                },
                                                                "biography": {
                                                                    "type": [
                                                                        "string",
                                                                        "null"
                                                                    ]
                                                                },
                                                                "photo": {
                                                                    "type": [
                                                                        "object",
                                                                        "null"
                                                                    ],
                                                                    "properties": {
                                                                        "sm": {
                                                                            "type": "string"
                                                                        },
                                                                        "md": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "required": [
                                                                        "sm",
                                                                        "md"
                                                                    ]
                                                                },
                                                                "position": {
                                                                    "type": "integer"
                                                                },
                                                                "created_at": {
                                                                    "type": "string"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": [
                                                                "id",
                                                                "ballot",
                                                                "party",
                                                                "first_name",
                                                                "last_name",
                                                                "sex",
                                                                "slogan",
                                                                "biography",
                                                                "photo",
                                                                "position",
                                                                "created_at",
                                                                "updated_at"
                                                            ]
                                                        }
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "title",
                                                    "description",
                                                    "position",
                                                    "seats",
                                                    "allow_blank",
                                                    "candidates_count",
                                                    "candidates",
                                                    "created_at",
                                                    "updated_at"
                                                ]
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "required": [
                                                "page",
                                                "per_page",
                                                "total"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/ballots/{ballot}/candidates/order": {
            "put": {
                "operationId": "candidate.reorder",
                "description": "`candidates` must be exactly the UUIDs of the ballot's candidates, each once, in the\nwanted order (422 `candidates: set_mismatch` otherwise). A draft election only. Answers\nthe candidates, in the new order. Owner or manager. Limited to 240 requests per hour per user.",
                "summary": "Put the candidates of a ballot in order",
                "tags": [
                    "Candidate"
                ],
                "parameters": [
                    {
                        "name": "ballot",
                        "in": "path",
                        "required": true,
                        "description": "The ballot UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReorderCandidatesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "ballot": {
                                                        "type": "string"
                                                    },
                                                    "party": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "sex": {
                                                        "type": "string",
                                                        "enum": [
                                                            "male",
                                                            "female"
                                                        ]
                                                    },
                                                    "slogan": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "biography": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "photo": {
                                                        "type": [
                                                            "object",
                                                            "null"
                                                        ],
                                                        "properties": {
                                                            "sm": {
                                                                "type": "string"
                                                            },
                                                            "md": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "required": [
                                                            "sm",
                                                            "md"
                                                        ]
                                                    },
                                                    "position": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "ballot",
                                                    "party",
                                                    "first_name",
                                                    "last_name",
                                                    "sex",
                                                    "slogan",
                                                    "biography",
                                                    "photo",
                                                    "position",
                                                    "created_at",
                                                    "updated_at"
                                                ]
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/invitations/{invitation}": {
            "delete": {
                "operationId": "invitation.destroy",
                "description": "Its link stops working; the row is deleted. An invitation that is unknown, already\naccepted, already cancelled or another institution's answers 404. Owner only.",
                "summary": "Cancel an invitation",
                "tags": [
                    "Invitation"
                ],
                "parameters": [
                    {
                        "name": "invitation",
                        "in": "path",
                        "required": true,
                        "description": "The invitation UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/users/{user}": {
            "delete": {
                "operationId": "user.destroy",
                "description": "The user can no longer sign in and their next request is refused. Their address is freed;\ntheir name is kept for the audit log. An institution keeps at least one owner. Removing\noneself also ends one's own session. Owner only.",
                "summary": "Remove a user",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "description": "The user UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReorderBallotsRequest": {
                "type": "object",
                "description": "docs/api/ballots/PUT-elections-{election}-ballots-order.md.",
                "properties": {
                    "ballots": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "minItems": 1
                    }
                },
                "required": [
                    "ballots"
                ],
                "title": "ReorderBallotsRequest"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}